+++ hobj/learning_models/linear_learning_model/representation.py
